feat: 전투 스킬과 속성 효과 시스템 확장

+++ Assets/Scripts/Enemy/Enemy Components/EnemyAttack/FlyingShoot.cs
@@ -48,7 +48,7 @@ namespace Enemy
 
             if (_ctx.HasMultiAttackModules)
             {
-                _ctx.anim.SetFloat(AnimHashes.AttackIndex, _animIndex);
+                _ctx.anim.SetInteger(AnimHashes.AttackIndex, _animIndex);
             }
         }
